【退出登录警示框】-cxg

+++ src/components/Home.vue
@@ -79,7 +79,19 @@ export default {
     },
     methods: {
         //退出登录
-        logout() {
+        async logout() {
+            const confirmResult = await this.$confirm('此操作将退出登录, 是否继续?', '提示', //第一个字符串是内容，第二个是标题
+            {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+            }).catch(err => err)//箭头函数里只有一行代码，所以简写
+            //如果用户取消删除，则返回值为字符串 cancel
+            //如果用户确认删除，则返回值为字符串 confirm
+            console.log(confirmResult);
+             if(confirmResult !== 'confirm') {
+                return this.$message.info('已取消退出登录')
+            }
             window.sessionStorage.clear();//清空token
             this.$router.push('/login');//重定向到登录页
             this.$message.success('退出成功！');
